Added Render Interface and World Class
Render interface: 
Implementing it will allow for a class to be rendered by registering it in the world method "addRenderable" (see below)

World class
Stores all renderable objects in an arraylist "renderObjects"
Contains the render loop. All objects stored in the arraylist "renderObjects" will be rendered within the loop. Also redraws the background prior to this.
To register a object to be rendered call "world.addRenderable(Renderer renderer)"